+++ changelog.md
@@ -1,4 +1,8 @@
 
+**<span style="color:#56adda">0.0.3</span>**
+- Ensure comchap and comcut scripts are executable
+- Add option to limit files based on extension
+
 **<span style="color:#56adda">0.0.2</span>**
 - Add comchap and comcut to plugin